Fix create organization route layout
