+++ src/router/Router.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import { useAuthRoutes } from './routes/auth'
+import { useAuthRoutes, useCreateOrganizationRoutes } from './routes/auth'
 import { useDashboardRoutes } from './routes/dashboard'
 import { useProcessCreateRoutes } from './routes/process-create'
 import { useRootRoutes } from './routes/root'
@@ -9,8 +9,9 @@ export const RoutesProvider = () => {
   const auth = useAuthRoutes()
   const processCreate = useProcessCreateRoutes()
   const dashboard = useDashboardRoutes()
+  const createOrganizationRoute = useCreateOrganizationRoutes()
 
-  const router = createBrowserRouter([root, auth, processCreate, dashboard])
+  const router = createBrowserRouter([root, auth, processCreate, dashboard, createOrganizationRoute])
 
   return <RouterProvider router={router} />
 }
